refactor: split test file per component, add duplicate label test
TagPill tests stay in TagPill.integration.test.tsx, TagList tests
move to TagList.integration.test.tsx. Adds test verifying onAdd is
not called when typing a duplicate label.
